Move error processing into core
