+++ corehq/motech/generic_inbound/core.py
@@ -1,15 +1,55 @@
 from django.utils.translation import gettext as _
 
-from corehq.apps.hqcase.api.core import serialize_case
+from corehq.apps.hqcase.api.core import (
+    SubmissionError,
+    UserError,
+    serialize_case,
+)
 from corehq.apps.hqcase.api.updates import handle_case_update
+from corehq.apps.userreports.exceptions import BadSpecError
 from corehq.motech.generic_inbound.exceptions import (
     GenericInboundApiError,
     GenericInboundRequestFiltered,
     GenericInboundValidationError,
 )
+from corehq.motech.generic_inbound.utils import ApiResponse
 
 
-def execute_generic_api(domain, couch_user, device_id, context, api_model):
+def execute_generic_api(api_model, request_data):
+    try:
+        response_json = _execute_generic_api(
+            request_data.domain,
+            request_data.couch_user,
+            request_data.user_agent,
+            request_data.to_context(),
+            api_model,
+        )
+    except BadSpecError as e:
+        return ApiResponse(status=500, json={'error': str(e)})
+    except UserError as e:
+        return ApiResponse(status=400, json={'error': str(e)})
+    except GenericInboundRequestFiltered:
+        return ApiResponse(status=204)
+    except GenericInboundValidationError as e:
+        return _get_validation_error_response(e.errors)
+    except GenericInboundApiError as e:
+        return ApiResponse(status=500, json={'error': str(e)})
+    except SubmissionError as e:
+        return ApiResponse(status=400, json={
+            'error': str(e),
+            'form_id': e.form_id,
+        })
+    return ApiResponse(status=200, json=response_json)
+
+
+def _get_validation_error_response(errors):
+    return ApiResponse(status=400, json={
+        'error': 'validation error',
+        'errors': [error['message'] for error in errors],
+    })
+
+
+def _execute_generic_api(domain, couch_user, device_id, context, api_model):
     _apply_api_filter(api_model, context)
     _validate_api_request(api_model, context)
